// src/shared/types/geo/point.ts
export type LatLntPoint = {
  lat: number;
  lng: number;
};

export type LongitudeLatitudePoint = {
  longitude: number;
  latitude: number;
};

export type ExtremeCoordinatesLatLng = {
  topLatitude: number;
  topLongitude: number;
  bottomLatitude: number;
  bottomLongitude: number;
};

export type ExtremeCoordinatesType = {
  topLatitude: number;
  topLongitude: number;
  bottomLatitude: number;
  bottomLongitude: number;
};